Improved Marketplace panel page UI
