+++ components/marketplaceSubnav/stylesheet.scss
@@ -1,0 +1,3 @@
+.nfd-root .nfd-marketplace-sidebar a:focus {
+	--tw-ring-color : rgb(25 107 222) !important;
+}
